added lots of output information
A fully functional --help for omb is available.

Split the output into more columns dtype | method.

Fixes some minor problems with the grainsize,
now taskloop constructs can use grainsize by
extracting the grainsize from OMP_SCHEDULE.
It is not pretty, but for this benchmark
it works simple and effictive.
Note that is almost always a good idea
for grainsize specification for taskloop.
The default of 1 is *HORRIBLE*!

Signed-off-by: Nick Papior <nickpapior@gmail.com>