Added system_clock information and tick
Now the user can see the precision of the timing.
We still rely on omp_get_wtime for consistency.
But essentially we *could* go back to the system_clock
timing methods.
However, GCC says that the precision of the rate can be
either micro or nano seconds, without really knowing which it is.
So it will be a system dependent variable, which is a annoying.
On my box wtick shows a ns precision. So that should be fine for
most cases. Only when an OpenMP runtime shows micro seconds should
one be careful since for small array sizes will this have bad
cases.

Signed-off-by: Nick Papior <nickpapior@gmail.com>