Made INT_TYPE to an option, and clarified bit-sizes in code
Now -DOMB_INT_TYPE=int32|int64 can be used to set the size
of loop counters etc. When using int32 the program can
stop when the number of array elements exceeds the int32 limit.

Further clarified the column output when doing --help.
They are now clearly explained with units.

Changed the data-type output to be in bytes. This seems more
natural, and less obscure in terms of string choices for later
data-type additions.
So now the dtype column contains the array element size in bytes.
A warning will be created if a data-type is not in full bytes.
I.e. storage_size % 8 != 0.

Added a new column with first touch information, 0 or 1.

Signed-off-by: Nick Papior <nickpapior@gmail.com>