added a serial method for checking
Also added a -test argument to only run
1 iteration, and printout the resulting
sum of the a array.
This means that the initialization of the arrays
required to be changed a bit so they are
different.
So I *guess* the simple check is sufficient.

Added a stop when the selected kernel is not
implemented.

Also, moved all OpenMP environment printing
into a separate module.
This enables simpler structure to re-use
certain routines for printing blocks of information.
In particular the place information can now be
processed depending on teams, and threads etc.

Fixed a bug for std. calculation for it == 1.

Signed-off-by: Nick Papior <nickpapior@gmail.com>