use masked if available, changed taskloop
Now we explicitly add the shared clause for taskloop.
We omit the nowait on the single. In principle we
could still use the nowait, since the omp barrier
has an implicit task-scheduling point.

Changed the grainsize option for taskloop to be a
number of tasks instead.
The default is now that it chooses the minimum number of
task based on:

- number of elements / 8
- 10 * number of threads

In this way we should get better perf. for lower
thread counts.

Added note on teams:distribute. The reason it performs
to bad is because each team created is still not shifting
the placement of their threads. I.e. all main team threads
will place their cores on the first place encountered
in OMP_PLACES.

Signed-off-by: Nick Papior <nickpapior@gmail.com>