implemented default(none) everywhere
This makes it more verbose, but forces a stricter
implementation. It can always be discussed whether
this is necessary (it isn't), but it also helps
checking compiler understanding of the semantics
in the omp-pragmas.

Also made the timing a bit simpler, for immediate
storage of the timing in an array element.

Signed-off-by: Nick Papior <nickpapior@gmail.com>