added quad + quad-fs method (4 array triad)
This should stream even more and possibly
show fma4 for AMD machines.
Intel machines only use fma3.

Also added debug information to the driver.
And cleaned that bash script so it is simpler.

The false-sharing is mainly to see how bad it
performs under various large mem access patterns.

Also, ensured that the kernel for the false-sharing
methods are always do. We might look into using
the teams construct which could work in the same way.

Signed-off-by: Nick Papior <nickpapior@gmail.com>