simplified fypp parsing
Added VERBATIM to add_custom_command which does
not require double escaping strings for fypp arguments.

Changed discovery of TEAMS, DEVICE, PARTITION
so that it is a simpler definition check.

Added a check_compiler directory with a simple
code structure that simply checks if the compiler
can compile the things.

Added more checks to the OpenMP constructs at
configure time.

Signed-off-by: Nick Papior <nickpapior@gmail.com>