cleaned omb-driver and enabled bigger domains
Now it can handle both threads and cores.

Ensured that when requesting the -env, it won't run
the benchmark. No need to wait for that.

Automatic handling of multiple domains and places.
If a place encompass multiple domains, it will automatically
be duplicated in the places list.
This might require other finetunings.

Signed-off-by: Nick Papior <nickpapior@gmail.com>