restructured data accumulation
The data layout is simpler and more direct.

Now the results_t contains a benchmark info type
that holds information on the benchmark it self.
It should likely be moved somewhere else, but
it makes sense to get out that information from
the results type.

This also means that certain data is filled at definition
time, and not at run-time.
This makes things a bit more robust.

Moved dryrun to the options type.